replace CustomStorageService with OfflineCachePersistenceService across the codebase
transition the persistence management to the new OfflineCachePersistenceService, removing deprecated methods from CustomStorageService
