+++ src/entrypoints/background/messages/unlockExtension.ts
@@ -1,7 +1,6 @@
 import ExtensionSettingsService, { ExtensionSettingsOptions } from "~/services/ExtensionSettingsService";
 import PassmanClientService from "~/services/PassmanClientService";
 import ExtensionUnlockService from "~/services/ExtensionUnlockService";
-import CustomStorageService from "~/services/CustomStorageService";
 import { onMessage } from '../messaging';
 import { logger } from "~/services/ConsoleLoggingService";
 
@@ -27,7 +26,6 @@ onMessage('unlockExtension', async (message) => {
 
         if (status && message.data.refreshAfterUnlock) {
             try {
-                await CustomStorageService.ensureExtensionPassmanClientPersistenceService();
                 const backendPassmanClient = await PassmanClientService.getBackendPassmanClient();
                 logger.debug("backendPassmanClient", backendPassmanClient);
